risk_mitigation/risk_model.py
import numpy as np
import scipy.optimize as sco

class RiskModel:
    def __init__(self, data, risk_free_rate):
        self.data = data
        self.risk_free_rate = risk_free_rate

    def calculate_returns(self):
        # Calculate the returns based on the provided data
        returns = np.diff(self.data) / self.data[:-1]
        return returns

    def calculate_volatility(self):
        # Calculate the volatility based on the returns
        returns = self.calculate_returns()
        volatility = np.std(returns)
        return volatility

    def calculate_sharpe_ratio(self, returns):
        # Calculate the Sharpe Ratio based on the returns and risk-free rate
        excess_returns = returns - self.risk_free_rate
        sharpe_ratio = np.mean(excess_returns) / np.std(returns)
        return sharpe_ratio

    def calculate_portfolio_returns(self, weights):
        # Calculate the portfolio returns based on the weights and returns
        returns = self.calculate_returns()
        portfolio_returns = np.dot(returns, weights)
        return portfolio_returns

    def calculate_portfolio_volatility(self, weights):
        # Calculate the portfolio volatility based on the weights and returns
        returns = self.calculate_returns()
        portfolio_volatility = np.sqrt(np.dot(weights.T, np.dot(np.cov(returns), weights)))
        return portfolio_volatility

    def optimize_portfolio(self):
        # Perform portfolio optimization to maximize the Sharpe Ratio

        returns = self.calculate_returns()

        def objective_function(weights):
            return -self.calculate_sharpe_ratio(returns)  # Maximizing Sharpe Ratio

        num_assets = len(returns)
        initial_weights = np.ones(num_assets) / num_assets  # Equal weights as initial guess

        constraints = (
            {'type': 'eq', 'fun': lambda weights: np.sum(weights) - 1}  # Constraint: weights sum to 1
        )

        bounds = tuple((0, 1) for _ in range(num_assets))  # Bounds: 0 <= weights <= 1

        optimization_result = sco.minimize(
            objective_function,
            initial_weights,
            method='SLSQP',
            bounds=bounds,
            constraints=constraints
        )

        optimized_weights = optimization_result.x
        return optimized_weights

    def analyze_risk(self):
        # Perform risk analysis using historical data and optimization

        returns = self.calculate_returns()
        volatility = self.calculate_volatility()
        sharpe_ratio = self.calculate_sharpe_ratio(returns)
        optimized_weights = self.optimize_portfolio()

        return {
            "returns": returns,
            "volatility": volatility,
            "sharpe_ratio": sharpe_ratio,
            "optimized_weights": optimized_weights
        }

    def mitigate_risk(self, risk_analysis):
        # Develop a risk mitigation plan based on the analysis results and optimized weights

        optimized_weights = risk_analysis["optimized_weights"]

        # Implement risk mitigation strategies based on the optimized weights
        # For example, you can apply rebalancing, hedging techniques, or diversification strategies

        mitigation_plan = {
            "optimized_weights": optimized_weights,
            # Add specific mitigation actions or strategies
        }
        return mitigation_plan
